move methods to new trait in primitive_sealed
